stompclt 1.2 (new formula)
stompclt is a versatile tool to interact with messaging brokers speaking
STOMP and/or message queues (see Messaging::Message::Queue) on disk.

It receives messages (see Messaging::Message) from an incoming module,
optionally massaging them (i.e. filtering and/or modifying), and sends
them to an outgoing module. Depending on which modules are used, the
tool can perform different operations.